requirements.txt
Pillow==9.1.0
python-decouple==3.6
requests==2.27.1
Pyrebase4==4.5.0